+++ crates/index/src/kd_tree/mod.rs
@@ -1,0 +1,5 @@
+pub mod index;
+pub mod types;
+
+#[cfg(test)]
+mod tests;
